requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
Django==3.2.15
django-ninja==1.4.3
idna==3.10
jsonschema==4.24.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
mdutils==1.7.0
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
openapi3-parser==1.1.21
packaging==25.0
pathable==0.4.4
prance==25.4.8.0
pydantic==2.11.7
pydantic_core==2.33.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rpds-py==0.25.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
